# clinic-reservation-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Clinic Reservation API",
        "description": "API documentation for the clinic reservation system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://clinic-reservation-api.test",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new patient or doctor",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "role"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "Hossam Mahmoud"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "hossam@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "01012345678",
                                        "nullable": true
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "PATIENT",
                                        "enum": [
                                            "PATIENT",
                                            "DOCTOR"
                                        ]
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "MALE",
                                        "nullable": true,
                                        "enum": [
                                            "MALE",
                                            "FEMALE"
                                        ]
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1999-01-15",
                                        "nullable": true
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Cairo",
                                        "nullable": true
                                    },
                                    "medical_history": {
                                        "type": "string",
                                        "example": "No chronic diseases",
                                        "nullable": true
                                    },
                                    "specialty": {
                                        "type": "string",
                                        "example": "Dermatology",
                                        "nullable": true
                                    },
                                    "available_time": {
                                        "type": "string",
                                        "example": "5PM - 9PM",
                                        "nullable": true
                                    },
                                    "bio": {
                                        "type": "string",
                                        "example": "Senior dermatologist",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User registered successfully"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "1|abcxyztoken"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "full_name": {
                                                            "type": "string",
                                                            "example": "Ahmed Hassan"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "format": "email",
                                                            "example": "ahmed@example.com"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "01012345678",
                                                            "nullable": true
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "PATIENT"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error / user already exists"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ahmed@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful"
                                        },
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9..."
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "full_name": {
                                                            "type": "string",
                                                            "example": "Ahmed Hassan"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "format": "email",
                                                            "example": "ahmed@example.com"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "01012345678",
                                                            "nullable": true
                                                        },
                                                        "role": {
                                                            "type": "string",
                                                            "example": "PATIENT"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}